Implement order service functionality and update dependencies
- Added new Makefile targets for database management and service setup.
- Updated Go module dependencies and removed unused ones.
- Introduced new database migration scripts for orders and order products.
- Implemented gRPC handlers for order creation and retrieval.
- Refactored order service to support pagination and improved error handling.
- Added support for generating Go code from proto files.

+++ internal/database/db/models.go
@@ -12,7 +12,7 @@ type Order struct {
 	ID          int32            `json:"id"`
 	UserID      int32            `json:"user_id"`
 	Status      string           `json:"status"`
-	TotalAmount int32            `json:"total_amount"`
+	TotalAmount float64          `json:"total_amount"`
 	CreatedAt   pgtype.Timestamp `json:"created_at"`
 	UpdatedAt   pgtype.Timestamp `json:"updated_at"`
 }
@@ -22,7 +22,7 @@ type OrderProduct struct {
 	OrderID   int32            `json:"order_id"`
 	ProductID int32            `json:"product_id"`
 	Quantity  int32            `json:"quantity"`
-	Price     int32            `json:"price"`
+	Price     float64          `json:"price"`
 	CreatedAt pgtype.Timestamp `json:"created_at"`
 	UpdatedAt pgtype.Timestamp `json:"updated_at"`
 }
